+++ src/aura-helper.js
@@ -19,6 +19,7 @@ Commands.Metadata.LocalRetrieveSpecial.createCommand(program);
 Commands.Metadata.OrgList.createCommand(program);
 Commands.Metadata.OrgDescribe.createCommand(program);
 Commands.Metadata.OrgCompare.createCommand(program);
+Commands.Metadata.OrgBetweenCompare.createCommand(program);
 Commands.Metadata.OrgRetrieveSpecial.createCommand(program);
 Commands.Metadata.OrgPermissions.createCommand(program);