+++ src/commands/metadata/index.js
@@ -9,5 +9,6 @@ exports.LocalRetrieveSpecial = require('./localRetrieveSpecial');
 exports.OrgList = require('./orgList');
 exports.OrgDescribe = require('./orgDescribe');
 exports.OrgCompare = require('./orgCompare');
+exports.OrgBetweenCompare = require('./orgBetweenCompare');
 exports.OrgRetrieveSpecial = require('./orgRetrieveSpecial');
 exports.OrgPermissions = require('./orgPermissions');