+++ src/commands/metadata/localDescribe.js
@@ -16,7 +16,7 @@ let argsList = [
     "root",
     "all",
     "type",
-    "sendTo",
+    "outputFile",
     "progress",
     "beautify"
 ];
@@ -29,8 +29,8 @@ exports.createCommand = function (program) {
         .option('-a, --all', 'Describe all metadata types stored in your local project')
         .option('-t, --type <MetadataTypeNames>', 'Describe the specified metadata types. You can select a single metadata or a list separated by commas. This option does not take effect if you choose describe all')
         .option('-p, --progress <format>', 'Option for report the command progress. Available formats: ' + CommandUtils.getProgressAvailableTypes().join(','))
-        .option('-s, --send-to <path/to/output/file>', 'Path to file for redirect the output')
         .option('-b, --beautify', 'Option for draw the output with colors. Green for Successfull, Blue for progress, Yellow for Warnings and Red for Errors. Only recomended for work with terminals (CMD, Bash, Power Shell...)')
+        .option('--output-file <path/to/output/file>', 'Path to file for redirect the output')
         .action(function (args) {
             run(args);
         });
@@ -48,11 +48,11 @@ async function run(args) {
         Output.Printer.printError(Response.error(ErrorCodes.FILE_ERROR, 'Wrong --root path. Select a valid path'));
         return;
     }
-    if (args.sendTo) {
+    if (args.outputFile) {
         try {
-            args.sendTo = Paths.getAbsolutePath(args.sendTo);
+            args.outputFile = Paths.getAbsolutePath(args.outputFile);
         } catch (error) {
-            Output.Printer.printError(Response.error(ErrorCodes.FILE_ERROR, 'Wrong --send-to path. Select a valid path'));
+            Output.Printer.printError(Response.error(ErrorCodes.FILE_ERROR, 'Wrong --output-file path. Select a valid path'));
             return;
         }
     }
@@ -71,12 +71,12 @@ async function run(args) {
         return;
     }
     describeLocalMetadata(args).then(function (result) {
-        if (args.sendTo) {
-            let baseDir = Paths.getFolderPath(args.sendTo);
+        if (args.outputFile) {
+            let baseDir = Paths.getFolderPath(args.outputFile);
             if (!FileChecker.isExists(baseDir))
                 FileWriter.createFolderSync(baseDir);
-            FileWriter.createFileSync(args.sendTo, JSON.stringify(result, null, 2));
-            Output.Printer.printSuccess(Response.success("Output saved in: " + args.sendTo));
+            FileWriter.createFileSync(args.outputFile, JSON.stringify(result, null, 2));
+            Output.Printer.printSuccess(Response.success("Output saved in: " + args.outputFile));
         } else {
             Output.Printer.printSuccess(Response.success("Describe Metadata Types finished successfully", result));
         }